:earth_asia: Location_info Code refactor
