+++ src/main/java/com/trip/IronBird_Server/map/adapter/controller/LocationController.java
@@ -2,7 +2,7 @@ package com.trip.IronBird_Server.map.adapter.controller;
 
 import com.trip.IronBird_Server.map.adapter.dto.LocationRequestDto;
 import com.trip.IronBird_Server.map.adapter.dto.LocationResponseDto;
-import com.trip.IronBird_Server.map.application.LocationService;
+import com.trip.IronBird_Server.map.application.LocationServicelmp;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,11 +15,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class LocationController {
 
-    private final LocationService locationService;
+    private final LocationServicelmp locationServicelmp;
 
     @PostMapping("/reverse-geo")
     public ResponseEntity<LocationResponseDto> reverseGeo(@RequestBody LocationRequestDto requestDto){
-        LocationResponseDto result = locationService.reverseGeoCoding(requestDto);
+        LocationResponseDto result = locationServicelmp.reverseGeoCoding(requestDto);
 
         return ResponseEntity.ok(result);
     }
