+++ src/main/java/com/trip/IronBird_Server/map/infrastructure/TourRepository.java
@@ -1,0 +1,7 @@
+package com.trip.IronBird_Server.map.infrastructure;
+
+import com.trip.IronBird_Server.map.domain.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TourRepository extends JpaRepository<Location, Long> {
+}
